+++ src/core/util/stringconversion.cpp
@@ -53,7 +53,7 @@ std::vector<std::string> splitString(const std::string& str, char delimeter) {
 
 std::string joinString(const std::vector<std::string>& str, std::string delimeter /*= ' '*/) {
     std::stringstream ss;
-    for(int i = 0; i < str.size()-1; ++i){
+    for(size_t i = 0; i < str.size()-1; ++i){
         ss << str[i] << delimeter;
     }
     ss << str.back();
